requirements.txt
asgiref==3.5.2
backoff==2.2.1
beautifulsoup4==4.11.1
branca==0.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
colorama==0.4.6
comtypes==1.1.14
cryptography==38.0.4
dj-database-url==1.2.0
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
dlib==19.18.0
face-recognition==1.3.0
face-recognition-models==0.3.0
feedparser==6.0.10
Flask==2.2.2
folium==0.14.0
html2text==2020.1.16
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.23.5
opencage==2.1.0
opencv-python==4.6.0.66
phonenumbers==8.13.2
Pillow==9.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyjokes==0.6.0
PyMsgBox==1.0.9
pyOpenSSL==22.1.0
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.7
pywakit== 0.0.5
pywin32==301
realpython-reader==1.1.1
requests==2.28.1
sgmllib3k==1.0.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
sqlparse==0.4.3
tomli==2.0.1
torch==1.13.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
Werkzeug==2.2.2
wikipedia==1.4.0
word2number==1.1
